fix: keep safecast

+++ packages/strategies/src/Accountant.sol
@@ -176,8 +176,8 @@ contract Accountant is ReentrancyGuardTransient, Ownable2Step {
         PackedVault storage _vault = vaults[msg.sender];
         uint256 vaultSupplyAndIntegral = _vault.supplyAndIntegralAndPendingRewardsSlot;
 
-        uint256 supply = uint96(vaultSupplyAndIntegral & StorageMasks.SUPPLY_MASK);
-        uint256 integral = uint96((vaultSupplyAndIntegral & StorageMasks.INTEGRAL_MASK) >> 96);
+        uint256 supply = vaultSupplyAndIntegral & StorageMasks.SUPPLY_MASK;
+        uint256 integral = (vaultSupplyAndIntegral & StorageMasks.INTEGRAL_MASK) >> 96;
 
         // Process any pending rewards if they exist and there is supply
         if (pendingRewards > 0 && supply > 0) {
@@ -236,10 +236,9 @@ contract Accountant is ReentrancyGuardTransient, Ownable2Step {
         PackedAccount storage _account = accounts[vault][account];
         uint256 accountBalanceAndRewards = _account.balanceAndRewardsSlot;
 
-        uint256 balance = uint96(accountBalanceAndRewards & StorageMasks.BALANCE_MASK);
-        uint256 accountIntegral = uint96((accountBalanceAndRewards & StorageMasks.ACCOUNT_INTEGRAL_MASK) >> 96);
-        uint256 accountPendingRewards =
-            uint64((accountBalanceAndRewards & StorageMasks.ACCOUNT_PENDING_REWARDS_MASK) >> 192);
+        uint256 balance = accountBalanceAndRewards & StorageMasks.BALANCE_MASK;
+        uint256 accountIntegral = (accountBalanceAndRewards & StorageMasks.ACCOUNT_INTEGRAL_MASK) >> 96;
+        uint256 accountPendingRewards = (accountBalanceAndRewards & StorageMasks.ACCOUNT_PENDING_REWARDS_MASK) >> 192;
 
         // Update pending rewards based on the integral difference
         accountPendingRewards += (currentIntegral - accountIntegral).mulDiv(balance, SCALING_FACTOR).toUint64();
@@ -299,14 +298,13 @@ contract Accountant is ReentrancyGuardTransient, Ownable2Step {
 
             // Unpack vault data
             vaultSupplyAndIntegral = _vault.supplyAndIntegralAndPendingRewardsSlot;
-            integral = uint96((vaultSupplyAndIntegral & StorageMasks.INTEGRAL_MASK) >> 96);
+            integral = (vaultSupplyAndIntegral & StorageMasks.INTEGRAL_MASK) >> 96;
 
             // Unpack account data
             accountBalanceAndRewards = _account.balanceAndRewardsSlot;
-            balance = uint96(accountBalanceAndRewards & StorageMasks.BALANCE_MASK);
-            accountIntegral = uint96((accountBalanceAndRewards & StorageMasks.ACCOUNT_INTEGRAL_MASK) >> 96);
-            accountPendingRewards =
-                uint64((accountBalanceAndRewards & StorageMasks.ACCOUNT_PENDING_REWARDS_MASK) >> 192);
+            balance = accountBalanceAndRewards & StorageMasks.BALANCE_MASK;
+            accountIntegral = (accountBalanceAndRewards & StorageMasks.ACCOUNT_INTEGRAL_MASK) >> 96;
+            accountPendingRewards = (accountBalanceAndRewards & StorageMasks.ACCOUNT_PENDING_REWARDS_MASK) >> 192;
 
             // Add new rewards if integral has increased
             if (integral > accountIntegral) {
@@ -408,7 +406,7 @@ contract Accountant is ReentrancyGuardTransient, Ownable2Step {
         uint256 donationAndIntegralTimestamp = _donation.donationAndIntegralTimestampSlot;
 
         // Add to existing donation amount
-        uint96 donation = uint96(donationAndIntegralTimestamp & StorageMasks.DONATION_MASK);
+        uint256 donation = donationAndIntegralTimestamp & StorageMasks.DONATION_MASK;
         donation += globalPendingRewards;
 
         // Store donation with current harvest integral
@@ -441,7 +439,7 @@ contract Accountant is ReentrancyGuardTransient, Ownable2Step {
         require(globalHarvestIntegral >= integral, HarvestIntegralNotReached());
 
         // Verify donation exists
-        uint256 donation = uint96(donationAndIntegralTimestamp & StorageMasks.DONATION_MASK);
+        uint256 donation = donationAndIntegralTimestamp & StorageMasks.DONATION_MASK;
         require(donation != 0, NoDonation());
 
         // Get donation premium percent
@@ -475,7 +473,7 @@ contract Accountant is ReentrancyGuardTransient, Ownable2Step {
         uint256 donationAndIntegralTimestamp = donations[account].donationAndIntegralTimestampSlot;
 
         // Get original donation amount
-        donation = uint96(donationAndIntegralTimestamp & StorageMasks.DONATION_MASK);
+        donation = donationAndIntegralTimestamp & StorageMasks.DONATION_MASK;
 
         // Get donation premium percent
         uint256 donationPremiumPercent = donations[account].donationPremiumPercent;
